Fixes issues with lznt1 extraction
The problem was not with lznt1 decompression but with reading the correct file size from $DATA attribute for files where $ATTRIBUTE_LIST was used. Some files will extract out with "invalid chunk header" in the console/log, but that is fine. Many files have empty chunks/blank data that exhibits this behavior. This happens because of the way compressed files are handled here, the sparse blocks from disk are written out as zeroed data, which when sent to the decompressor results in the "invalid chunk" message.